Multi page website
One-Page Website 

The goal of this project is a simple one-page website for a chosen television series, film, or game. It contains sections on plot, key info, cast/creative team, characters, and reviews. There is a navigation bar with anchor links that let users quickly jump to different sections. Styling is done using an external CSS file and we have at least two CSS classes. Furthermore, a button link with a styled link takes users to the official website in a new tab. 

Multi-Page Website (Bonus - Optional) 

This version builds on the one-page website by separating the content to multiple pages for better organization. Each of the sections — plot, key details, cast, characters and reviews — has its own dedicated page in a “pages” subdirectory. There is a consistent navigation bar for easy browsing on every page. The external CSS makes all pages look the same.