Three images + Styling those divs
Three Images 

In this project I present three images that are legally allowed for commercial use, all of which are related by a common theme. The alt attributes of each image is also appropriate for accessibility and the width and height attributes are used to keep the original aspect ratio. An additional context is provided in the form of a title attribute. The appearance of the images are modified using CSS filters and all of the images optimized to be below 300KB for performance. Copyright license details are displayed below each image. 

Bonus - Optional Styling Those Divs 

It is a task of changing the given starter code to match the reference design. CSS is used to style the images and surrounding divs creatively while keeping responsive. The final layout is neat and with proper spacing, alignment and visual effects.