Strc
Structuring a Webpage 

The main idea of this project is recreating the structural layout of a reference image provided to you, using proper semantic HTML elements. The webpage consists of structured header, navigation menu, main content (“Welcome” section), “Line Up” section with two images and footer. The fonts and color scheme that are specified in the external CSS file are used to apply them and the layout is close to the reference image. 

Study Planner (Bonus - Optional) 

This is a mockup of a single page web application for managing study plan of a Web Development module. On the page layout, there is a header, navigation bar, main content which is divided into 2 or 3 columns, and a footer. It contains five demo study items, which are presented in the structured format. It includes a button element and proper sematic HTML elements to make it clear and accessible. The layout is styled with an external CSS file that has more to do with design and structure than with interactivity. Also, commented code is included for better readability.