Embedded video + Alan Partridge soundboard
Embed Video 

The embedding of a locally hosted video into an HTML document with specific requirements is what this project is. The video loops, includes a relevant poster image before playback, and is centered both horizontally and vertically on the page with CSS. Good semantic HTML tags are accessible and it is clean. 

Alan Partridge Soundboard with Advanced Features 

  

This Alan Partridge soundboard has been enhanced to include: 

- Text to Speech Feature: This lets you type text that is immediately converted to audio and played. 

 

- Audio samples: It shows audio samples in 9 pages, with left and right arrows for moving through the pages. When appropriate (i.e.: left on the first page, right on the last page), the arrows are hidden. 

 

-Each audio sample now shows its length in seconds below its title.