Petrol Calculator + RGB Coloring game
Petrol Calculator 

The objective of the project is to build a petrol calculator that takes the cost per liter and the number of liters purchased as input and then calculates and displays the total cost. There is a default cost of 1.72 for petrol per liter, a default of 0 liters, and a button to calculate. The button has hover effects and input fields have focus effects, all of which has clean and professional styling along with comments throughout the code. 

RGB Colour Guessing Game (Bonus - Optional) 

In this interactive game, users are given an RGB value to guess and are given multiple color options. On clicking, the user has to guess the color and the game continues until there is no more lives left. The player will replay at the end and see the final score. The project is using index.html, style.css, and index.js with great comments and a great professional styling.